Move old tools to their own directory
