+++ EuPathDB/src/main/java/org/gusdb/oauth2/eupathdb/subscriptions/tools/SqlGenerator.java
@@ -1,8 +1,10 @@
-package org.gusdb.oauth2.eupathdb.tools;
+package org.gusdb.oauth2.eupathdb.subscriptions.tools;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.nio.charset.StandardCharsets;
+
+import org.gusdb.oauth2.eupathdb.tools.SubscriptionTokenGenerator;
 
 public class SqlGenerator {
